Ajoute limitation anti-force-brute, corrige un crash JWT critique, et deux retours utilisateur
- throttle:5,1 sur connexion/inscription/mot de passe oublie (aucune
  limite avant, force brute possible)
- config/jwt.php : cast (int) obligatoire sur JWT_TTL/JWT_REFRESH_TTL
  - sans ca, Carbon 3.x refuse la chaine renvoyee par env() et TOUTE
    connexion echoue en 500 des que JWT_TTL est defini explicitement
    (introduit par la reduction de duree de vie du jeton, 150 jours
    par defaut ramenes a 24h + renouvellement periodique cote front)
- Sidebar redimensionnable a la souris (glisser le bord droit),
  largeur memorisee entre sessions
- Panneau d'apercu de document elargi (288px -> 480px) - le cadre
  etait trop petit pour vraiment voir le contenu

+++ backend/config/jwt.php
@@ -101,7 +101,11 @@ return [
     |
     */
 
-    'ttl' => env('JWT_TTL', 60*3600),
+    // (int) obligatoire : env() renvoie toujours une chaîne depuis le .env, et
+    // Carbon 3.x (contrairement aux versions précédentes) n'accepte plus une
+    // chaîne dans addMinutes() — sans ce cast, toute connexion échoue en 500
+    // dès que JWT_TTL est défini explicitement dans le .env.
+    'ttl' => (int) env('JWT_TTL', 60*3600),
 
     /*
     |--------------------------------------------------------------------------
@@ -120,7 +124,8 @@ return [
     |
     */
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    // Même raison que 'ttl' ci-dessus : cast obligatoire pour Carbon 3.x.
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
 
     /*
     |--------------------------------------------------------------------------
